requirements.txt
streamlit
pandas
requests
PyPDF2
python-docx
openpyxl
sentence-transformers
google-generativeai
qdrant-client
